requirements.txt
certifi==*
charset-normalizer==*
idna==*
PyYAML==*
requests==*
urllib3==*